backend/timetable-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlmodel==0.0.19
python-jose==3.3.0
